# std::io: Use sendfile for UnixStream