@nicokoch 2020-12-04
@the8472 I also noticed CopyRead and CopyWrite are not implemented for `Box<T> where T: Copy{Read|Write}`. Is this deliberate or an oversight?