@the8472 2020-12-04
I assumed that is handled by auto-deref. If it isn't then it should be added.